osql: send preprocessed blobs to the master.
Instead of sending a blob to the master as is, pack and compress the blob on a replicant, and then send these preprocessed blob to the master.
This approach reduces the network IO between replicants and the master, as well as the memory footprint and CPU use of the master.