Fix for live vutf8 schema change
This is an edge case when extending inline size of a vutf8 field:

Let P and Q denote the inline size in the old and new schemas, respectively.
Stripe S finishes converting. A new record now lands on stripe S. It must be
written to both old and new tables as the schema change hasn't completed yet.
Let X denote the length of the vutf8 field in the new record.

If `P < X <= Q`, when `osql_odh_blob` is on, the insert would fail because
`utf8_validate()` would be examining wrong data and the schema change would
subsequently abort; when it is off, the database would crash in an invalid
`free`.

The bug was introduced by blob preprocessing (#1399). The fix is simple:
If the old or new type of a field is vutf8, we unodhfy the payload of any new
records during the schema change.

(DRQS 155195897)

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>