Fix pmux on WSL
Unix Socket on Windows 10 doesn't support `SO_KEEPALIVE`. `evconnlistener_new_bind()`
skips `SO_KEEPALIVE` if `_WIN32` is define. However on WSL the macro isn't predefined.

To work it around, we create a socket on our own, attempt to set `KEEPALIVE` on it
but do not bail out if failed, and pass it to `evconnlistener_new()` to listen on.

Fixes #2290.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>