Fixing client invalid response type error
Morgan demonstrated that if SSL is enabled, when the server experiences
a high connection rate and is slow to complete an SSL handshake,
a client may end up reading an incomplete response type 369296128 (0x16030300),
which is the header of an OpenSSL's hello packet.

This patch ensures that a socket is closed if its SSL handshake fails.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>